remove export of include directories
Currently no headers are being installed. When building a workspace with `--isolated` this fails the build of downstream packages trying to use this package.

As an alternative the headers could be installed and the export could stay as it is. Since the headers were not used anymore I just removed the export.